added: data for about me section + placeholders for book section
