+++ frontend/components/layout.tsx
@@ -33,7 +33,8 @@ export default function Layout({ children, title, page }) {
         <meta property="og:url" content="https://joseizarra.com/" />
 
       </Head>
-      { page !== 'CV' && <NavBar page={page} />}
+      {/* { page !== 'CV' && <NavBar page={page} />} */}
+      <NavBar page={page} />
       <main className={cn("min-h-screen  font-sans antialiased", fontSans.variable)}>
           {children}
           <Analytics />
