+++ frontend/json/navbarlinks.json
@@ -17,6 +17,6 @@
     },
     {
         "name": "Thoughts",
-        "href": "/thoughts"
+        "href": "/thoughts/home"
     }
 ]
