--- frontend/pages/thoughts.tsx
@@ -1,9 +1,0 @@
-import Layout from '../components/layout';
-import UnderConstruction from '../components/underConstruction';
-
-
-export default function Thoughts() {
-    return (
-        <UnderConstruction />
-    )
-}
