+++ frontend/pages/thoughts/home.tsx
@@ -1,0 +1,24 @@
+import Layout from '../../components/layout';
+import UnderConstruction from '../../components/underConstruction';
+
+
+
+//  requires custom layout for the thoughts section
+
+export default function Thoughts() {
+    return (
+        <Layout page="Thoughts" title="Thoughts | Jose Izarra">
+            <h1></h1>
+            
+            {/* Link to some of my articles + philosophy */}
+            <article>
+
+            </article>
+
+            {/* Book Section */}
+            <article>
+
+            </article>
+        </Layout>
+    )
+}
